remove warning

+++ app/rust/src/parser/certificate/hash_tree.rs
@@ -295,9 +295,9 @@ impl<'a> TryFrom<&RawValue<'a>> for HashTree<'a> {
         cfg_if::cfg_if! {
             if #[cfg(all(not(test), not(feature = "clippy"), not(feature = "fuzzing")))] {
                 let f = unsafe {
-                    let raw_fn_ptr = Decoder::new as *const () as u32; // Convert to raw address
-                    let adjusted_ptr = crate::pic_addr(raw_fn_ptr); // Apply PIC offset
-                    core::mem::transmute::<u32, fn(&[u8]) -> Decoder>(adjusted_ptr) // Convert back to fn type
+                    let raw_fn_ptr = Decoder::new as *const () as usize; // Use usize instead of u32
+                    let adjusted_ptr = crate::pic_addr(raw_fn_ptr as u32) as usize; // Cast back to usize after PIC adjustment
+                    core::mem::transmute::<usize, fn(&[u8]) -> Decoder>(adjusted_ptr) // Use usize for transmute
                 };
 
                 let mut d = f(value.0);
